lesson_02-2/requirements.txt
fastapi==0.110.1
fastavro==1.9.4
httpx==0.27.0
pydantic==2.6.4
pytest-asyncio==0.23.6
pytest-mock==3.14.0
requests==2.31.0
uvicorn==0.29.0
pyyaml==6.0.1
pandas==2.2.2
